pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshai"
version = "0.1.0"
description = "LLM-powered Meshtastic mesh network assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "K7ZVX", email = "matt@echo6.co"}
]
keywords = ["meshtastic", "llm", "mesh", "lora", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
]

dependencies = [
    "meshtastic>=2.3.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.4.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
meshai = "meshai.main:main"

[project.urls]
Homepage = "https://github.com/zvx-echo6/meshai"
Repository = "https://github.com/zvx-echo6/meshai"

[tool.setuptools.packages.find]
where = ["."]
include = ["meshai*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
